perf(hook): dispatch agent-hook without constructing the command surface
ESM static imports are evaluated before the importing module's own body runs, so every
top-level import in the process entry is paid on every invocation whatever was asked for.
With the whole command surface in `src/index.ts`, that meant the MCP server, the viewer,
the tree-sitter code indexer, the config UI, the AI pipeline and the transcript indexer
were all loaded before commander had looked at argv.

Affordable for a command a human types. Not affordable for `agent-hook`, which is a fresh
process per agent tool call -- hundreds a session -- and needs the store, the host-hook
normaliser and nothing else.

So `src/index.ts` is now a dispatcher: `agent-hook` goes straight to a new
`src/cli/agent-hook.ts`, everything else dynamic-imports the command surface, which moves
verbatim to `src/cli/program.ts`. `--help` falls through to the program, so
`knowl agent-hook --help` still describes the command instead of waiting on stdin.
`catchUpTranscripts` becomes a dynamic import inside the one branch that reaches it.

No behaviour change: the hook body is the same code, and `knowl agent-hook` remains a
registered command.

Measured on this machine, interleaved A/B of the two real bundles, 21 calls each, run
twice: 410ms -> 269ms and 478ms -> 317ms per hook call. 34% both times, ~7s and ~11s per
200-call session. (Absolutes are inflated by seven agents sharing the machine; the ratio
is what held.) dist/index.js goes from 522,718 bytes to 670.

The guard is structural rather than wall-clock: a timing assertion on a loaded machine
measures the machine, while one static import of `./cli/program.js` put back into the
entry would restore the entire cost silently.
